requirements.txt
flask==3.0.3
flask-limiter==3.8.0
scikit-learn==1.5.2
xgboost==2.1.3
pandas==2.2.3
numpy==1.26.4
joblib==1.4.2
python-whois==0.9.5
requests==2.32.3
tldextract==5.1.3
matplotlib==3.9.4
seaborn==0.13.2
streamlit==1.40.2
plotly==5.24.1
colorama==0.4.6
python-dotenv==1.0.1
gunicorn==23.0.0